refactor: move status label and date formatting functions to utility files
